fixup fake exchange rate manager for integration tests
